refactor(cli): extract fix.py command module from cli monolith
Second slice of the cli monolith decomposition. Moves the fix command,
its three helpers (_collect_target_files, _print_fix_result,
_run_interactive_fix), and its 20 local constants out of
phi_scan/cli/__init__.py into phi_scan/cli/fix.py.

The fix_command function is imported into __init__.py and registered via
app.command("fix")(fix_command) to preserve the user-facing CLI. No
behavior change; no test surface change (no tests import fix internals).

cli/__init__.py shrinks 1285 → 1135 lines.
